Create _.compact and pass specs
The problem with _.map is that is creates a 1 to 1 relationship between
function input and output. In the case of creating _.compact, a 1 to 1
relationship is undesirable. In such instances _.filter is the better
choice since it allows you to filter the contents of the final array

'!!' or 'not not' allows you to cast any variable to it's boolean
equivalent