Create _.without and pass spec
The line '_.without = (array, values...) ->' uses CoffeeScripts splat
operator (...). The splat operator replaces Javascripts arguments
array-like object. The splat operator is used here to assign 'values' as
the container which will hold all additional arguments.

The line 'value in values' takes advantage of CoffeeScripts contains
operator (in). 'in' checks for the existence of a given value inside a
provided container.