Create _.zip and pass specs
This method uses _.sortBy to find the array with the greatest length. We
need the array with the greatest length to handle the edge case that not
all arrays passed in are the same size. If this is the case, we would
need to keep zipping the values of the longest array with undefined
values in place of the short arrays.

Originally a while loop was going to be used, but incrementing a count
proved too cumbersome so I decided to just use a _.map call. Hence, the
outter _.map call is collecting all the zipped arrays, and the inner
_.map call is zipping the inner arrays together.