Create _.intersection and pass spec
Using _.compact with _.map makes it easy to filter values. The problem
with _.filter is that you cannot specify what you want to be returned.
It returns the values that pass the predicate. Sometimes you want a
transformed value returned if the predicate is passed. In instances like
this it is best to use _.compact in conjunction with _.map