Create _.sortedIndex and pass specs
Lessons Learned:

An Iteratee could be either a callback or a property. It is anything
which allows you to transform a given value. _.transform is a helper
method which determines what Iteratee is and performs the correct action
based on that to transform the provided value.

When creating callbacks that do not return a boolean value, remember to
have an else that returns an alternative value. Many callbacks are
predicates, which do not require creating an else statement since
returning an expression can evaluate to only true or false. Remember
when returning non-boolean values to include alternative return values.