Create _.range and pass specs
Final notes on finishing up 'arrays' section:

I was disappointed I could not get rid of the line 'if start is stop
then start = 0' in the _.range function. I tried using the comma operator as follows: 'stop = stop or start, start = 0' but I couldn't get the syntax right. It seems strange, but I have seen a very similar looking line of code in the book "If Hemwingway wrote Javascript".

My primary objective in designing these functions is experimentation and code readability. It is difficult to make these functions readable because they are abstract which makes naming difficult. These functions are not concerned with a concrete implementations. These functions are concerned with abstracting out commonly used programming techniques. The nature of the challenges at hand makes for stimulating experimentation.