Create _.bind and pass specs
_.bind changes the context of a function without executing the function
unlike call or apply. Additionally, _.bind can applly arguments to the function in advance - 'partial application'.

_.bind achieves partial application by taking
advantage of Javascript closures. _.bind creates a wrapper function
which is returned to it's caller. This wrapper function behaves just like the original function passed _.bind except it has arguments and context prefilled.

The arguments are passed during runtime to the wrapper function and then are
merged with the partial arguments passed to _.bind. The merged arguments along
with the context specified in the _.bind header are all passed to
fn.apply which results in the function call.