Create _.debounce and test manually
_.debounce accepts a function and wait period as arguments. It then returns a
new function which is essentially the function that was passed into _.debounce with a wrapper that provides additional functionality. My version of _.debounce makes it so every call to the debounced function during
the wait period, resets that wait period. Every call to the function
when the wait period is not active, calls the function once and then
initiates the wait period.