Create _.wrapper and pass spec
'args.unshift(fn)' prepends the function to the array. Normally we would
concatenate the arguments from the wrapper function with the arguments
from the anonymous function, but since we are simply applying a single
value from the wrapper we will just prepend it to the anonymous
function's arguments using unshift.

When using apply to call a function without changing the context
set the first argument in the apply method to the function being called.
The apply method is actually nothing but a wrapper for the call method.
Apply simply breaks down the arguments array into comma seperated form.