Create _.compose and pass spec
_.compose records the results from each function call. The
results are then used as arguments for the next function call. Since apply
is being used each function call needs to be wrapped in an array:

args = [fn.apply(fn, args)]

This way args can easily be passed to apply in order to execute the next function call.

There are two specific reasons CoffeeScripts splat
operator (...) is being used rather than JavaScripts builtin arguments pseudo-array.
Firstly, if the arguments pseudo-array were being used there would need to be a variable to hold
arguments from the wrapper function, if we wanted to refer to arguments
from the outer function body. When function body's are nested,
'arguments' refers to the formal parameters of the current function.
Using splat operators makes it easy to access arguments from
different functions when dealing with nested functions. This means that the
arguments pseudo-array will not be shadowed when their are
nested functions.

Secondly, using splat operator transforms the arguments
psuedo-array into a real array. This is what the splat operator is
doing under the hood:

functions = 1 <= arguments.length ? __slice.call(arguments, 0) : [];

The arguments pseudo-array has slice applied to it which copies over the format
parameters into a real array. Using a real array instead of a pseudo-array
provides the ability to use methods which would be inaccessible other.