Create _.invert and pass spec
I chose to use delete here for memory efficiency (and because I have
been looking for a reason to use delete). the delete operator removes
the reference to a property on an object. It does not actually remove
the value itself, it only removes the reference- from their the garbage
collector deallocates the value from memory when it detects that value
no longer has a reference to it.

If I did not use delete when creating _.invert then I would of had to
created a new objects to hold the results. Instead of using delete, as I
do in the following function:

_.invert = (obj) ->
  _.each obj, (value, key) ->
    obj[value] = key
    delete obj[key]
  obj

I would of had to create a new object to store the values, like this:

_.invert = (obj) ->
  result = {}
  _.each obj, (value, key) ->
    result[value] = key
  result