Create _.isFunction and _.pick and pass specs
_.pick is the equivalent of _.filter, but for objects instead of arrays.
Like _.filter, _.pick accepts a container and a predicate callback, and
returns a new container based on what values the predicate evaluates to
true.

Alternatively, instead of passing in a predicate, _.pick also
accepts a list of keys which acts as a white list. Whatever keys are on
the list will determine what keys will be included in the return
container.