Create all type comparison functions
The isFinite function already exist on the global object. The point of recreating isFinite is the
underscore version accounts for the fact that the builtin version does
not properly process falsy values. The built in isFinite function returns
true for all falsy values through type coercion.

For example:
isFinite('') == true

because '' coerces to zero and zero is a finite number. The underscore
version I created accounts for this mistake.

_.isNaN = (value) ->
  Number.isNaN(value)

The builtin isNaN has a similar issue as the builtin isFinite. The
builtin isNaN returns true is passed the value undefined, which is
clearly correct. _.isNaN uses the isNaN method on the Number object,
which only returns true if indeed the input is NaN.

I originally thought 'not a number' meant a mathematical expression evaluated to a value that was a type other than Number. I later realized that 'not a number' really means 'not a real number'. NaN is actually meant to represent mathematical expressions which result in values that cannot be expressed in the number set JavaScript provides. For example, imaginary numbers cannot be represented in JavaScript. If an expression is performed that creates an imaginary number then JavaScript will raise NaN. Imaginary numbers are numbers though, they are just numbers that cannot be represented in JavaScript. Therefore, 'not a number' should more accurately be called 'not a real number'. JavaScript can only represent real numbers and everything else is considered NaN. This is why the data type of NaN is number.

typeof NaN
>>>"number"